requirements.txt
openai
pdfplumber
pandas
tabulate
chromadb
python-dotenv
streamlit
tiktoken
